requirements.txt
backports.csv==1.0.7
biopython==1.74
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.4
django-filter==2.2.0
django-import-export==1.2.0
django-nested-admin==3.2.3
django-widget-tweaks==1.4.5
et-xmlfile==1.0.1
eutils==0.5.2
idna==2.8
jdcal==1.4.1
lxml==4.4.1
mysqlclient==1.4.4
numpy==1.17.1
odfpy==1.4.0
openpyxl==2.6.2
pkg-resources==0.0.0
python-monkey-business==1.0.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
urllib3==1.25.3
xlrd==1.2.0
xlwt==1.3.0
